fix(registry) values should not be weak
both keys and values in the registry of already generated Kong
objects were weak. This works for single value entities, but
the `servers` entity has two values. So in that case an
intermediate table is created, but that intermediate table is
not anchored anywhere (only the key, and the two entries inside
the table are anchored). Hence in some cases the garbage collector
kicked in and removed the intermediate entry prematurely.

This caused a 'nil'-reference in a later pass requesting the same
entry from the registry.